requirements.txt
altair
annotated-types
anyio
appnope
asttokens
attrs
blinker
cachetools
certifi
charset-normalizer
click
comm
debugpy
decorator
distro
executing
gitdb
GitPython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
idna
ipykernel
ipython
jedi
Jinja2
jiter
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
markdown-it-py
MarkupSafe
matplotlib-inline
mdurl
narwhals
nest-asyncio
numpy
openai
packaging
pandas
parso
pexpect
pillow
platformdirs
prompt_toolkit
proto-plus
protobuf
psutil
ptyprocess
pure_eval
pyarrow
pyasn1
pyasn1_modules
pydantic
pydantic_core
pydeck
Pygments
pyparsing
python-dateutil
python-dotenv
pytz
pyzmq
referencing
requests
rich
rpds-py
rsa
six
smmap
sniffio
stack-data
streamlit
tenacity
toml
tornado
tqdm
traitlets
typing_extensions
tzdata
uritemplate
urllib3
wcwidth
